fix: use type long long for n=100000000

+++ task3.c
@@ -3,14 +3,14 @@
 
 int main() {
     double partial_sum, total_sum;
-    double n = 100000000;
+    long long n = 100000000;
 
     #pragma omp parallel private(partial_sum) shared(total_sum) {
         partial_sum = 0;
         total_sum = 0;
 
         #pragma omp for
-        for (int i = 0; i <= n; i++) {
+        for (long long i = 0; i <= n; i++) {
             double numerator = 1;
             if (i%2 == 1) {
                 numerator *= -1;
